LPS-179177 Extract CommerceCatalogAccountGroupsScreenNavigationCategory methods to a CommerceCatalogAccountGroupsScreenNavigationEntry component and extends CommerceCatalogAccountGroupsScreenNavigationCategory in it
